Setup basic file structure for the application
The file strucutre should be structured in a relatively intuitive way.
Each new component will have a .js & .scss file as well as a test file.
In addition, I configured storybook for testing the visual state of the
components within the application. They will be in their own file as
they get tested in isolation from the rest of the app.